fix: resolve MyPy errors in checkpoint.py and lmstudio_provider.py
- Add type ignores for checkpoint manager attributes not in UIAgentProtocol
- Fix unused type ignores in lmstudio_provider.py unreachable statements
- Add proper type ignore for stream method override
- Fix CompletionResponse call with all required arguments

+++ victor/framework/graph_cache.py
@@ -364,7 +364,7 @@ class CompiledGraphCache:
             stats = self._stats.copy()
 
             total = stats["hits"] + stats["misses"]
-            stats["hit_rate"] = int(stats["hits"] / total if total > 0 else 0.0)
+            stats["hit_rate"] = stats["hits"] / total if total > 0 else 0.0
 
             if self._cache is not None:
                 stats["current_size"] = len(self._cache)
